refactor(sync): move the session-budget cluster out of the orchestrator
The renderer-heap measurement, the chunk-boundary pause verdict, the
post-preview prognosis, the run-start baseline and the QAM status payload
are about Steam's memory, not the mechanics of a sync. They now live in
SessionBudgetMonitor, a library sub-service beside the orchestrator over
the same LibrarySyncStateBox, and the orchestrator reaches them through
the SessionBudgetGate protocol instead of holding the renderer seams
itself. The pure arithmetic stays in domain/session_budget.py.

Behaviour is unchanged: same readings, same pause verdicts, same
callable payload, same fail-open degradation when the renderer cannot be
read. The module-size allowlist entry for the orchestrator drops from
1480 to its real new size.
