+++ py_modules/services/protocols/__init__.py
@@ -56,6 +56,7 @@ from services.protocols.cross_service import (
     SaveSortChangeFn,
     SaveSortingProvider,
     SessionAchievementSync,
+    SessionBudgetGate,
     SessionMigrationReader,
     SessionPlaytimeRecorder,
     SessionPostExitSync,
@@ -233,6 +234,7 @@ __all__ = [
     "SaveSortChangeFn",
     "SaveSortingProvider",
     "SessionAchievementSync",
+    "SessionBudgetGate",
     "SessionMigrationReader",
     "SessionPlaytimeRecorder",
     "SessionPostExitSync",
