+++ tests/services/test_disc_bake_sites.py
@@ -26,8 +26,6 @@ from unittest.mock import MagicMock
 import pytest
 from fakes.fake_active_core_resolver import FakeActiveCoreResolver
 from fakes.fake_disc_resolver import FakeDiscResolver
-from fakes.fake_renderer_gc import FakeRendererGc
-from fakes.fake_renderer_rss import FakeRendererRss
 from fakes.fake_unit_of_work import FakeUnitOfWork, FakeUnitOfWorkFactory
 from fakes.system_time import FakeClock, FakeSleeper, FakeUuidGen
 
@@ -121,8 +119,7 @@ class TestSyncOrchestratorBakeSite:
                 artwork=MagicMock(),
                 active_core=FakeActiveCoreResolver(default=(None, None)),
                 disc_resolver=disc_resolver,
-                renderer_rss=FakeRendererRss(),
-                renderer_gc=FakeRendererGc(),
+                session_budget=MagicMock(),
             )
         )
